Update Table.java
I changed the condition((Line 76)) in the for loop under the method "getColumnsNamesAsArray" from '<' to '<=' because the method was ignoring the last column name for the table. In order to see the last column name you had to ad an empty column name and empty row content on that column.